Moved GetNoteIdsWithAtLeastInterval() into CardRepository.
