+++ AnkiSentenceCardBuilder/DataAccessLayer/Repositories/CardRepository.cs
@@ -42,5 +42,14 @@ namespace AnkiJapaneseFlashcardManager.DataAccessLayer.Repositories
 			//Return success
 			return true;
 		}
+
+		public IEnumerable<long> GetNoteIdsWithAtLeastInterval(IEnumerable<long> noteIds, int interval)//Card(Note)
+		{
+			return _context.Cards
+						.Where(c => noteIds.Contains(c.NoteId))//Grab cards with matching note ids
+						.Where(c => c.Interval >= interval)//Filter cards with matching intervals
+						.Select(c => c.NoteId)//Grab the note ids
+						.ToList();//Return the list
+		}
 	}
 }
